Replaces the pneumatic with a servo
