+++ src/org/usfirst/frc/team4915/stronghold/RobotMap.java
@@ -2,7 +2,7 @@ package org.usfirst.frc.team4915.stronghold;
 
 import edu.wpi.first.wpilibj.CANTalon;
 import edu.wpi.first.wpilibj.DigitalInput;
-import edu.wpi.first.wpilibj.Solenoid;
+import edu.wpi.first.wpilibj.Servo;
 
 public class RobotMap {
     // If you are using multiple modules, make sure to define both the port
@@ -28,8 +28,8 @@ public class RobotMap {
     private static final int BOULDER_SWITCH_PORT = -1; // TODO
     private static final int LAUNCHER_BOTTOM_SWITCH_PORT = -1; // TODO
     private static final int LAUNCHER_TOP_SWITCH_PORT = -1; // TODO
-
-    private static final int LAUNCHER_SOLENOID_PORT = -1; // TODO
+    
+    private static final int LAUNCHER_SERVO_PORT = -1; //TODO
     // not actual port values
 
     public static CANTalon intakeLeftMotor = new CANTalon(INTAKE_LEFT_MOTOR_PORT);
@@ -39,6 +39,5 @@ public class RobotMap {
     public static DigitalInput boulderSwitch = new DigitalInput(BOULDER_SWITCH_PORT);
     public static DigitalInput launcherTopSwitch = new DigitalInput(LAUNCHER_TOP_SWITCH_PORT);
     public static DigitalInput launcherBottomSwitch = new DigitalInput(LAUNCHER_BOTTOM_SWITCH_PORT);
-
-    public static Solenoid launcherSolenoid = new Solenoid(LAUNCHER_SOLENOID_PORT);
+    public static Servo launcherServo = new Servo(LAUNCHER_SERVO_PORT);
 }
