+++ src/org/usfirst/frc/team4915/stronghold/commands/IntakeLauncher/ActivateLauncherServoCommand.java
@@ -1,0 +1,31 @@
+package org.usfirst.frc.team4915.stronghold.commands.IntakeLauncher;
+
+import edu.wpi.first.wpilibj.command.Command;
+import org.usfirst.frc.team4915.stronghold.Robot;
+
+public class ActivateLauncherServoCommand extends Command {
+
+    public ActivateLauncherServoCommand() {
+        requires(Robot.intakeLauncher);
+    }
+
+    protected void initialize() {
+        Robot.intakeLauncher.activateLaunchServo();
+    }
+
+    protected void execute() {
+    
+    }
+
+    protected boolean isFinished() {
+        return true;
+    }
+
+    protected void end() {
+
+    }
+
+    protected void interrupted() {
+    
+    }
+}
