+++ src/org/usfirst/frc/team4915/stronghold/commands/IntakeLauncher/RetractLauncherServoCommand.java
@@ -1,0 +1,30 @@
+package org.usfirst.frc.team4915.stronghold.commands.IntakeLauncher;
+
+import edu.wpi.first.wpilibj.command.Command;
+import org.usfirst.frc.team4915.stronghold.Robot;
+
+public class RetractLauncherServoCommand extends Command {
+
+    public RetractLauncherServoCommand() {
+        requires(Robot.intakeLauncher);
+    }
+
+    protected void initialize() {
+        Robot.intakeLauncher.retractLaunchServo();
+    }
+
+    protected void execute() {
+    
+    }
+
+    protected boolean isFinished() {
+        return true;
+    }
+
+    protected void end() {
+    }
+
+    protected void interrupted() {
+    
+    }
+}
